requirements.txt
appdirs
pyyaml
pyperclip
termcolor
colorama
pyqt5
